requirements/prod.txt
aiohttp>=3.7.4
asyncpg>=0.22.0
bugsnag>=4.0.3
deep-translator>=1.5.4
emoji>=1.4.2
redis>=3.5.3
tortoise-orm>=0.17.7
twitchio>=2.0.6